requirements.txt
transformers>=4.49.0
huggingface_hub>=0.23.1
pillow
bitsandbytes
torch 